move setup detail to vue template
